Color-only/hybrid mode impl. with fail-safes
- New color-only/hybrid mode
- New fail-safes (color and general)
- Param "min_num_midpts_color_correct" sets the amount of next midpoints
that have to have the correct color. Set it to 0 for color-blind mode.
Set it to infinite (big number) for color-only mode. Everything
in-between will be a hybrid mode, i.e. closer midpoints will fulfill
color constrains but color beyond "min_num_midpts_color_correct" will
not be considered.
- Param "allow_oneside" specifies that every midpoint beyond
"min_num_midpts_color_correct" will have at least one cone with correct
color.
- Param "max_way_horizon_size" specifies the max sight length, i.e. the
length of the path ahead of the car.
- Param "failsafe_max_way_horizon_size" specifies the max sight length
during a fail-safe event.
- Param "color_failsafe" allows the color fail-safe to be turned on.
Color-only mode will be deactivated and sight length set to
"failsafe_max_way_horizon_size".
- Param "general_failsafe_safetyFactor" sets the factor by which search
parameters will be increased during a general fail-safe event.
- Param "general_failsafe" allows the general fail-safe to be turned on.
Color-only mode will be deactivated, search parameters increased by
"general_failsafe_safetyFactor" and sight length set to
"failsafe_max_way_horizon_size".
- Readme gets updated with color-only/hybrid mode explanation,
fail-safes explanation and new result gif(s).

+++ src/structures/Edge.cpp
@@ -13,7 +13,7 @@
 /* ----------------------------- Private Methods ---------------------------- */
 
 uint64_t Edge::computeHash(const Node &n0, const Node &n1) {
-  return (std::max(n0.id, n1.id) << HASH_SHIFT_NUM) + std::min(n0.id, n1.id);
+  return (uint64_t(std::max(n0.id, n1.id)) << HASH_SHIFT_NUM) + std::min(n0.id, n1.id);
 }
 
 double Edge::computeLen(const Node &n0, const Node &n1) {
